Fix unitialized _nesting_level on new event-loop instances.
We currently monkey-patch class methods on the event loop, and check for
the existence of our monkey patches when deciding whether or not to
(re-)install them.  However, our implementation also depends on an
instance variable _nesting_level which is only instantiated on the
event-loop instance currently being patched.

This patch makes _nesting_level a class variable, effectively making all
new event-loop instances start with an initial _nesting_level of 0.