set loop back when loop end
When running task with `run_until_complete` we
expect that when the loop is done there is not running
loop any more, as mentioned in https://github.com/erdewit/nest_asyncio/pull/25
, and it the reason of many issues. However,
we can not simply set the running to None.

Here, after the new loop is done, the running loop
is set back to the original one.